Add scroll direction indicator.
Add the outer wrapper for indicator, add
indicator. Set height of indicator to 100vh,
and absolute position to bottom centre of
wrapper with translateY to 105% to hide
indicator below the wrapper. Add keyframe
to translateY by 105% and -105% right after
to hide the indicator at the top of wrapper.

Add right margin for scroll direction indicator
to media screen orientation to keep right margin
responsive to viewport.