Add animaions to nav bar
Create wrappers with overflow hidden, set
position of text to absolute and outide visible
mask so it is hidden and in position to animate.

Create key frame animation to translateY from
-105% to 0, where 0 is the position where text
is inside visible mask.

Add delays to animations for correct timing.

Add active id with higher opacity to text in
navBar that corresponds to current scene to show
which scene is active.