Add animated second page.
Creating a model of a visible scene on rails or
tracks. Where the visible scene remains in the
viewport while the 'tracks' scroll up and down
behind it. Allowing the user to scroll up and
down without scrolling the viewport's content.

Use scroll snap to allow page to scroll back and
forth with minimal scroll wheel movement. Use
sticky position with scenes class to make visible
scenes/pages stay in viewport while page scroll
snaps between the divs underneath scenes.

Use overflow hidden on scenes to hide scene2. Set
scene2 into position to animate in. Add scene2
animation to scene2.showing class in css.

Use js to get detect user scroll direction, use
as trigger to animate scene2 in and out.

In index.html change position of the #navBar to
outisde the #track so it stays on top of scene2.

Change #active from the nar bar using the scroll
direction as trigger to show which scene is
active.